feat: finishing registration screen
